Add detection of the WiX Toolset bin directory
Depending on the host platform, WiX Toolset might be installed in the "Program Files" or "Program Files (x86)" directory. This will check for each path in order to locate the directory instead of assuming that it exists. It will then use that directory as the location for the candle and light binaries.