Make update handling more robust
This change addresses the use case of a user installing versions of Auto OSSEC overtop of each other. Upgrades should go cleanly and remove previous versions. It also introduces a bugfix so that the EXECAUTOOSEC Custom Action is ran at a time when it has the auto_ossec binary available.